requirements.txt
azure-identity
azure-mgmt-containerservice
azure-containerregistry
pydantic
